make sort_xml_attrs a classmethod
